fib: increase the reference counter size for mfib to 32 bits
The 16 bits counter size limits the number of interfaces to 65K which is
too low for certain use cases. GTP-U for example creates a new interface
for every tunnel.

This change makes the mfib reference counter size inline with the fib one.

Type: fix

+++ src/vnet/mfib/mfib_table.h
@@ -99,7 +99,7 @@ typedef struct mfib_table_t_
     /**
      * number of locks on the table
      */
-    u16 mft_locks[MFIB_TABLE_N_LOCKS];
+    u32 mft_locks[MFIB_TABLE_N_LOCKS];
 
     /**
      * Table ID (hash key) for this FIB.
